cleaned up readme and some code, added template for .env and included axios in package.json which was missing

+++ src/Initializers/DigitalOcean/DigitalOceanInitializer.spec.ts
@@ -6,16 +6,19 @@ import { MACHINE_TYPES } from '../../constants';
 import type { NodeSSH } from 'node-ssh';
 
 const TEST_SSH_NAME = 'TEST_SSH_NAME'
-
+const TEST_DROPLET_NAME = 'ASAP-VPS-Test-Droplet';
 let initializer: DigitalOcean;
-jest.setTimeout(100000);
+jest.setTimeout(120000); // 2 minutes
 
 describe('DigitalOcean', () => {
   beforeAll(async () => {
-      initializer = new DigitalOcean(process.env.DIGITAL_OCEAN_API_KEY as string);
-      await initializer.getOptions(true, true);
-      await initializer.deleteKeysWithName(TEST_SSH_NAME);
-      await asyncTimeout(500);
+    if(!process.env.DIGITAL_OCEAN_API_KEY) {
+      throw new Error('DIGITAL_OCEAN_API_KEY is not set in the environment variables');
+    }
+    initializer = new DigitalOcean(process.env.DIGITAL_OCEAN_API_KEY as string);
+    await initializer.getOptions(true, true);
+    const deletedCount = await initializer.deleteKeysWithName(TEST_SSH_NAME);
+    await asyncTimeout(500);
   });
   beforeEach(async () => {
     expect(process.env.DIGITAL_OCEAN_API_KEY).toBeTruthy();
@@ -33,17 +36,22 @@ describe('DigitalOcean', () => {
     expect(created.privateKey).toBeDefined();
 
     // should have no test key
+    const existingKey = initializer.options!.ssh?.find(s => s.name === TEST_SSH_NAME);
     expect( initializer.options!.ssh?.find(s => s.name === TEST_SSH_NAME)).toBeUndefined();
+    
     const initialized = await initializer.addSSHKey(created.publicKey!, TEST_SSH_NAME)
     expect(initialized).toBeDefined();
     expect(initialized.fingerprint).toBeDefined();
     expect(initialized.name).toBeDefined();
     expect(initialized.id).toBeDefined();
     expect(initialized.publicKey).toBe(created.publicKey);
+    
     await initializer.getOptions(true, true);
     expect( initializer.options!.ssh?.find(s => s.name === TEST_SSH_NAME)).toBeDefined();
+    
     await initializer.deleteSSHKey(initialized.id)
     expect( initializer.options!.ssh?.find(s => s.name === TEST_SSH_NAME)).toBeUndefined();
+    
     await initializer.getOptions(true, true);
     expect( initializer.options!.ssh?.find(s => s.name === TEST_SSH_NAME)).toBeUndefined();
   });
@@ -51,10 +59,13 @@ describe('DigitalOcean', () => {
   it('tests DigitalOcean.createServer', async () => {
     const ssh = await SSHService.createKeys();
     expect(ssh?.publicKey).toBeDefined();
+    
     const initialized = await initializer.addSSHKey(ssh.publicKey!, TEST_SSH_NAME)
+    
     const createdDroplet = await initializer.createServer({
       region: 'nyc1',
       size: 's-1vcpu-1gb',
+      name: TEST_DROPLET_NAME,
       image: MACHINE_TYPES.UBUNTU_22,
       ssh: initialized.id
     });
@@ -66,16 +77,22 @@ describe('DigitalOcean', () => {
 
   it('tests DigitalOcean.createServer and can ssh into it.', async () => {
     const created = await SSHService.createKeys();
+    
     const initializedSsh = await initializer.addSSHKey(created.publicKey!, TEST_SSH_NAME)
+    
     const createdDroplet = await initializer.createServer({
       region: 'nyc1',
       size: 's-1vcpu-1gb',
       image: MACHINE_TYPES.UBUNTU_22,
-      ssh: initializedSsh.id
+      ssh: initializedSsh.id,
+      name: TEST_DROPLET_NAME
     });
-    const ssh : NodeSSH = await SSHService.connect(createdDroplet.ip, created.privateKey);
+    
+    const ssh : NodeSSH = await SSHService.connect(createdDroplet.ip, created.privateKey, undefined, {maxRetries: 12, retryTimeout: 10000});
+    
     expect(ssh).toBeDefined();
     await ssh.dispose();
+    
     await initializer.deleteSSHKey(initializedSsh.id)
     await initializer.deleteServer(createdDroplet.id);
   });
